app/api/main.py
import boto3
import os
import json
from google.cloud import secretmanager

def get_aws_credentials():
    # In a real scenario, you'd fetch these from GCP Secret Manager
    # For this skeleton, we'll assume they are available or mocked
    client = secretmanager.SecretManagerServiceClient()
    # TODO: Implement secret fetching logic
    return {
        'aws_access_key_id': 'TODO',
        'aws_secret_access_key': 'TODO'
    }

def main(request):
    """Responds to any HTTP request.
    Args:
        request (flask.Request): HTTP request object.
    Returns:
        The response text or any set of values that can be turned into a
        Response object using
        `make_response <http://flask.pocoo.org/docs/1.0/api/#flask.Flask.make_response>`.
    """
    
    # CORS headers
    if request.method == 'OPTIONS':
        headers = {
            'Access-Control-Allow-Origin': '*',
            'Access-Control-Allow-Methods': 'GET',
            'Access-Control-Allow-Headers': 'Content-Type',
            'Access-Control-Max-Age': '3600'
        }
        return ('', 204, headers)

    headers = {
        'Access-Control-Allow-Origin': '*'
    }

    try:
        # creds = get_aws_credentials()
        # dynamodb = boto3.resource('dynamodb', 
        #                           region_name='us-east-1',
        #                           aws_access_key_id=creds['aws_access_key_id'],
        #                           aws_secret_access_key=creds['aws_secret_access_key'])
        
        # table = dynamodb.Table('VisitorCount')
        # response = table.update_item(
        #     Key={'id': 'count'},
        #     UpdateExpression='ADD visits :inc',
        #     ExpressionAttributeValues={':inc': 1},
        #     ReturnValues="UPDATED_NEW"
        # )
        # count = response['Attributes']['visits']
        
        # Mock response for skeleton
        count = 42 
        
        return (json.dumps({'count': int(count)}), 200, headers)
        
    except Exception as e:
        return (json.dumps({'error': str(e)}), 500, headers)
