app/api/requirements.txt
boto3
google-cloud-secret-manager
functions-framework
